Split config.changed into changed and new
While it can be argued that config values are "changing" from not having
a previous value in the very first hook that's fired, not distinguishing
that from a "real" change (that is, the config had a definite value
previously and has a new definite value now) makes certain constructs
difficult.  On the other hand, it is relatively easy to go the other
way using @when_any.