Rebuild venv for post-series-upgrade hook
After a series upgrade, the venv needs to be completely rebuilt to
account for updates to the system's Python.  This PR also moves the
logic for upgrade-charm wheelhouse updates out of the hook code so that
that logic is all in one spot and so we can drop the specialized hook
code when the hook template improvements get into stable in the charm
snap.