Add python_packages layer option
There are cases where Python packages require compilation on the target
system architecture and thus can't be packaged in the wheelhouse.  This
allows them to be installed as part of the normal bootstrap process so
they don't need to be manually managed by the charm and do the right
thing with regard to the venv.