Add feature to disable all charm code execution
As a workaround for https://bugs.launchpad.net/juju/+bug/1741506 this
allows a charm author who needs to remove a unit which is in error state
to disable further charm code execution to make it possible to
(uncleanly) remove the unit despite any charm errors that are blocking
it.  The command to disable the charm is:

```
juju run unit/0 -- disable-charm
```

There's no way provided to re-enable, since this is intended as a last
resort.