Add version to charm-env and prevent downgrades
If two charms are co-located or subordinate with differing versions of
the charm-env helper it can cause issues if the older version gets
installed. This change adds a version number to the helper and logic
during the bootstrap to detect the installed version and upgrade it only
if its own version is newer. It will also re-do the check for every
hook, to handle the case where an older charm without this logic
overwrites it after the fact.

Fixes #136