Fix broken pip due to downgrade on Focal+
Focal provides a newer version of pip (and setuptools) than what is
bundled, and downgrading it in the venv leads to it being broken. This
doesn't show until after the first bootstrap, so on the first
upgrade-charm hook. This skips installing pip and setuptools if pip is
already at a newer version in the venv.

Fixes #159