Use ``travis_wait`` to keep func test alive
At present we have enabled excessive debug-logging to accompilsh
the same thing, but that makes it unpleasant to look at.

Change permissions on the LXD socket to give travis user access
instead of using the sudo back to ourself trick as it leads to
some unwanted complexity.

Add Python 3.8 target to test matrix.

Replace ``cryptography`` with ``psycopg2`` as it has not banned
Python 3.4 support.

Closes #161