Add LANG=C.UTF-8 if it isn't set for subprocesses
LANG is used to determine the default decoder when reading files in
Python. If it isn't set, some python versions simply default to ASCII
which leads to errors in some cases during Python package installation.

Juju did not set LANG in hook environments until 2.7 where it started to
add LANG=C.UTF-8. In order to support pre-2.7 environments where the
presence of LANG is significant during the package installation and
upgrades, let's add LANG explicitly if it is not set.

Fixes: #173