Fix upgrades on focal for charms built pre-160
Prior to PR #160, deploys on focal would downgrade pip and setuptools
and leave the venv in a broken state that would only reveal itself at
charm upgrade.  PR #160 prevents that broken state for new deploys, but
this fixes upgrading deploys that were already in that broken state.

Tested with the ubuntu charm by first confirming upgrade failure on both
a build pre-160 and post-160, then confirming that both cleanly upgrade
with this fix in place.