Fix upgrade flag file created too early
During an upgrade a flag file is created at `wheelhouse/.upgrade` to
indicate that a charm upgrade has been started so that when the
interpreter is reloaded after updating the venv, it doesn't get stuck in
a loop repeatedly doing the upgrade.  However, this flag file is created
before the upgrade is successfully completed, so a failed upgrade can
leave the flag file around and make the charm think the upgrade already
happened, even though it actually failed. This can leave the charm in a
state where it cannot retry the upgrade and is stuck in a broken state.

Fixes #189