Pin pip and setuptools
It would be wonderful if we could just have these unpinned and
use the latest and gratest, but unfortunately new releases
routinely bring incompatible and braking changes.

The consequence is released charms breaking after a rebuild, and
more often than you would think we actually encounter changes
in the dependency change between charm CI run and merge, leading
to upload of broken charms to charmhub, which is unacceptable.